Hotel Management
Talking about the features of this Simple Hotel Management System, at first, the user has to enter his/her data. It includes the name of the user, address, check-in, and check-out dates. The user can calculate room rents. Inside this section, there are total four types of room with different prices. After selecting the room type, the system asks to enter the number of nights spent in order to calculate room rent. This simple system also contains other functions such as calculating restaurant, laundry and game bill. When the user selects to calculate restaurant bill, the system displays a small menu. From there the user has to select foods and then it displays the total restaurant bill. The other remaining features; calculating laundry and game bill also follows the same procedure as of calculating restaurant bill.